requirements.txt
SQLAlchemy>=1.3
python-slugify~=4.0.1
